[no-issue] move: StartUp class module bootstrap to extension
