+++ java-apm-agent/java-agent-bootstrap/src/main/java/org/traffichunter/javaagent/bootstrap/AgentExecutionEngine.java
@@ -26,8 +26,6 @@ package org.traffichunter.javaagent.bootstrap;
 import java.io.File;
 import java.lang.instrument.Instrumentation;
 import java.lang.reflect.Constructor;
-import java.time.Duration;
-import java.time.Instant;
 import java.util.jar.JarFile;
 
 /**
@@ -57,8 +55,6 @@ public final class AgentExecutionEngine {
     }
 
     private void run() {
-        StartUp startUp = new StartUp();
-        Instant startTime = startUp.getStartTime();
 
         if(AgentExecutionEngine.class.getClassLoader() != null) {
             throw new IllegalStateException("AgentExecutionEngine is not loaded in bootstrap class loader");
@@ -80,55 +76,14 @@ public final class AgentExecutionEngine {
 
             TrafficHunterAgentStarter trafficHunterAgentStarter = startAgent(ClassLoader.getSystemClassLoader());
 
-            trafficHunterAgentStarter.start(inst, agentArgs, startTime);
+            trafficHunterAgentStarter.start(inst, agentArgs);
         } catch (Exception e) {
             throw new IllegalStateException("Failed to start agent", e);
         }
-
-        log.info("Started TrafficHunter Agent in {} second", startUp.getUpTime());
     }
 
     public static void run(final File agentBootstrapJar, final String args, final Instrumentation inst) {
         new AgentExecutionEngine(agentBootstrapJar, System.getProperty(args), inst).run();
-    }
-
-    /**
-     * The {@code StartUp} class extends {@link LifeCycle} to measure the agent's
-     * startup durations.
-     *
-     * <p>Features:</p>
-     * <ul>
-     *     <li>Tracks the agent's start time and end time.</li>
-     *     <li>Calculates the total uptime.</li>
-     * </ul>
-     */
-    private static class StartUp extends LifeCycle {
-
-        public StartUp() {
-            super();
-        }
-
-        @Override
-        public Instant getStartTime() {
-            return this.startTime;
-        }
-
-        @Override
-        public Instant getEndTime() {
-            if(endTime == null) {
-                this.endTime = Instant.now();
-            }
-            return endTime;
-        }
-
-        @Override
-        public Double getUpTime() {
-            if(getStartTime() == null && getEndTime() == null) {
-                throw new IllegalStateException("No start time or end time specified");
-            }
-
-            return Duration.between(getStartTime(), getEndTime()).toMillis() / 1_000.0;
-        }
     }
 
     // Agent starter invokes reflection
